+++ R/RcppExports.R
@@ -29,6 +29,10 @@ iterate_maze <- function(X, x, y) {
     .Call('_aRtsy_iterate_maze', PACKAGE = 'aRtsy', X, x, y)
 }
 
+iterate_phyllotaxis <- function(iter, a, p) {
+    .Call('_aRtsy_iterate_phyllotaxis', PACKAGE = 'aRtsy', iter, a, p)
+}
+
 draw_planet <- function(X, radius, xcenter, ycenter, threshold, iterations, starprob, ncolors, colorsused, fade, lightright) {
     .Call('_aRtsy_draw_planet', PACKAGE = 'aRtsy', X, radius, xcenter, ycenter, threshold, iterations, starprob, ncolors, colorsused, fade, lightright)
 }