+++ NEWS.md
@@ -5,6 +5,7 @@
 - Added `canvas_remacan()`. Thanks to @akident on Twitter for the suggestion.
 - Added `canvas_phyllotaxis()`.
 - Added `canvas_cobweb()`.
+- Added `canvas_chladni()`.
 
 # aRtsy 0.1.4